Fix error on creating new abs-file in Emacs
`abs--calculate-input-files` assumes that the `buffer-file-name` belongs to a
file that exists. This is not the case when opening a new abs-file that has not
yet been saved.

The function `abs-flymake-mode-on` simply checks if the file exists on the
file-system, and turns flymake on as usual if it does. If not, the function
adds itself to the `after-save-hook` (locally) so that it will be run once the
file is saved. The hook is always removed when flymake has successfully turned
on.